requirements.txt
absl-py==0.13.0
altgraph==0.16.1
aniso8601==9.0.1
arabicnlp==0.1.7
ArabicOcr==1.1.6
asgiref==3.2.10
astor==0.8.0
astroid==2.4.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.2.0
auto-py-to-exe==2.7.8
backcall==0.1.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.0
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
bottle==0.12.17
bottle-websocket==0.2.9
branca==0.4.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
clang==5.0
click==7.1.2
colorama==0.4.1
convertdate==2.3.2
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.14
dataclasses==0.8
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
djangorestframework==3.11.0
docutils==0.15.2
dse==3.3.0
dython==0.4.1
easyocr==1.4
Eel==0.12.4
efficientnet==1.1.1
emoji==0.5.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
entrypoints==0.3
ephem==4.0.0.1
fasttext==0.9.2
filelock==3.0.12
Flask==2.0.1
Flask-RESTful==0.3.9
flatbuffers==1.12
folium==0.10.1
fse==0.1.15
future==0.17.1
gast==0.4.0
gensim==3.8.3
geocoder==1.38.1
geographiclib==1.50
geopy==1.22.0
GetOldTweets3==0.0.11
gevent==1.4.0
gevent-websocket==0.10.1
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
hijri-converter==2.1.2
idna==2.8
imageio==2.9.0
importlib-metadata==0.23
importlib-resources==5.2.2
imutils==0.5.4
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.4.2
itsdangerous==2.0.1
jedi==0.15.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-tabnine==1.2.0
kanren==0.2.3
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kivy-deps.glew==0.1.12
kiwisolver==1.1.0
korean-lunar-calendar==0.2.1
langcodes==2.0.0
lazy-object-proxy==1.4.3
lxml==4.5.0
marisa-trie==0.7.5
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==1.0.0
multipledispatch==0.6.0
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.5.1
nltk==3.4.5
notebook==6.0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
Owlready2==0.26
packaging==20.4
pandas==1.1.5
pandocfilters==1.4.2
parso==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.2.1
plac==0.9.6
platformdirs==2.3.0
pluggy==0.13.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.17.3
psutil==5.7.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.9.0
PyArabic==0.6.8
pyaramorph==0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycparser==2.19
pygame==1.9.6
Pygments==2.4.2
PyInstaller==3.5
pylint==2.6.0
PyMeeus==0.5.11
pyparsing==2.4.2
PyPDF2==1.26.0
pyquery==1.4.3
pyrsistent==0.15.4
PySocks==1.7.1
pytesseract==0.3.8
pytest==5.4.3
python-bidi==0.4.2
python-crfsuite==0.9.7
python-dateutil==2.8.0
python-dotenv==0.19.0
pytz==2019.2
PyWavelets==1.1.1
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
ratelim==0.1.6
rdflib==5.0.0
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.3.1
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.94
setuptools-git==1.2
six==1.15.0
sklearn==0.0
sklearn-crfsuite==0.3.6
smart-open==2.1.0
soupsieve==2.0
spacy==2.1.8
sqlparse==0.3.1
srsly==0.1.0
tabulate==0.8.7
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
thinc==7.0.8
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
toolz==0.11.1
torch==1.7.0+cpu
torchaudio==0.7.0
torchsummary==1.5.1
torchtext==0.6.0
torchvision==0.8.1+cpu
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
tweepy==3.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
unification==0.2.2
urllib3==1.25.6
virtualenv==20.7.2
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0.1
whichcraft==0.6.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wordcloud==1.8.1
wordfreq==2.3.2
wrapt==1.12.1
xgboost==1.4.2
xlrd==1.2.0
zipp==3.5.0
